Refactor parts into separate files, rename
