+++ go-engine/doc/README.md
@@ -1,0 +1,12 @@
+# Development Notes
+
+## 2024-09-04
+
+Something like the following happened in the past few weeks:
+
+- Added a `Simulation` struct to manage the simulation state and parts.
+- Added a `Part` interface to represent a part in the simulation.
+- Added a `Boiler` struct to represent a boiler in the simulation.
+- Added a `Turbine` struct to represent a turbine in the simulation.
+- Added a `Generator` struct to represent a generator in the simulation.
+
