+++ go-engine/src/components/component.go
@@ -1,0 +1,9 @@
+package components
+
+import (
+	"won/sim-lab/go-engine/common"
+)
+
+type Component interface {
+	Update(env *common.Environment, otherComponents []Component)
+}
